requirements.txt
Django==2.1.7
djangorestframework==3.9.2
requests==2.20.1
mysqlclient==2.0.1
django-cors-headers==2.5.2
PyJWT==1.7.1
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.4.2
pytest